requirements.txt
alembic==1.4.0
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
elasticsearch==7.5.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
pendulum==2.0.5
pycodestyle==2.5.0
pycparser==2.19
pyes==0.99.6
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
pytzdata==2019.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
uliweb-apps==0.1.0
uliweb-layout==0.1.0
uliweb-ui==0.1.0
uliweb3==0.1
urllib3==1.25.8
wrapt==1.11.2
